backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database ORM
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9

# Date utilities
python-dateutil>=2.8.2

# Authentication
python-jose[cryptography]>=3.3.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AI & MCP
openai>=1.0.0
mcp>=1.0.0

# Server-Sent Events (SSE)
sse-starlette>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
